size

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	printf("size of a char:  %lu byte(s)\n", sizeof(char));
-	printf("size of an int:  %lu byte(s)\n",  sizeof(int));
-	printf("size of a long int:  %lu byte(s)\n",  sizeof(long int));
-	printf("size of a long long int:  %lu byte(s)\n", sizeof(long long int));
-	printf("size of a float:  %lu byte(s)\n", sizeof(float));
+	printf("size of a char:  %lu byte(s)\n", (int)sizeof(char));
+	printf("size of an int:  %lu byte(s)\n",  (int)sizeof(int));
+	printf("size of a long int:  %lu byte(s)\n",  (int)sizeof(long int));
+	printf("size of a long long int:  %lu byte(s)\n", (int)sizeof(long long int));
+	printf("size of a float:  %lu byte(s)\n", (int)sizeof(float));
 	return (0);
 }
